% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{elbow_method}
\alias{elbow_method}
\title{Plot a graph (number of clusters, total sum of squares within all the clusters created with the given method).}
\usage{
elbow_method(d, kmax = 10, method)
}
\arguments{
\item{d}{A data matrix.}

\item{kmax}{Maximal number of clusters to plot the clustering.}

\item{method}{Clustering method, either "kmeans" or "hclust".}
}
\value{
No return. Plots a graph.
}
\description{
Method must be either "kmeans" or "hclust". Default method, if none is given, is hclust.
}
\seealso{
\code{\link{kmeans_wss}}

\code{\link{hclust_wss}}
}
