% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polr_helpers.R
\name{invlogit}
\alias{invlogit}
\title{Inverse logit function.}
\usage{
invlogit(x)
}
\arguments{
\item{x}{A number, a vector of numbers or other numerical structure.}
}
\value{
Inverse logit, \code{exp(x)/(1+exp(x))}, with the same numerical strucure as \code{x}.
}
\description{
Inverse logit function.
}
