% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kmeans_wss}
\alias{kmeans_wss}
\title{Calculate total WSS for all the clusters created with kmeans method.}
\usage{
kmeans_wss(d, kmax = 10)
}
\arguments{
\item{d}{A data matrix.}

\item{kmax}{Maximal number of clusters to make the clustering.}
}
\value{
A vector of sum of squares within all the clusters for each number of clusters from 1 to kmax.
}
\description{
Calculate total WSS for all the clusters created with kmeans method.
}
\seealso{
\code{\link{wss.total}}
}
