% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polr_helpers.R
\name{toprobabilities}
\alias{toprobabilities}
\title{Transform cumulative probabilities (distribution) into single probabilities.}
\usage{
toprobabilities(cumprobs, nms = NULL)
}
\arguments{
\item{cumprobs}{A vector of cumulative probabilities (without the last 1).}

\item{nms}{Optional, names of the probabilities values.}
}
\value{
Vector of single probabilities.
}
\description{
Transform cumulative probabilities (distribution) into single probabilities.
Adds 1 at the end and calculates differences respectively.
}
\seealso{
\code{\link{invlogit}}
}
