% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{wss.cluster}
\alias{wss.cluster}
\title{Calculate total WSS within a cluster.}
\usage{
wss.cluster(clustermat)
}
\arguments{
\item{clustermat}{A data matrix representing cluster.}
}
\value{
Sum of squares within the cluster.
}
\description{
Calculates sum of square distances from the mean.
}
