% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{wss.total}
\alias{wss.total}
\title{Calculate total WSS for all the clusters – clustering set.}
\usage{
wss.total(dmatrix, labels)
}
\arguments{
\item{dmatrix}{A data matrix.}

\item{labels}{Clustering of the data matrix.}
}
\value{
Sum of squares within all the clusters.
}
\description{
Calculate total WSS for all the clusters – clustering set.
}
\seealso{
\code{\link{wss.cluster}}
}
